// frontend/src/app/components/model/labTest.model.ts
export interface LabTest {
  labTestId: number;
  adminId: number;
  test_name: string;
  status: string;
  description: string;
  test_time_start: string;
  test_time_end: string;
  delivery_time: string;
  life_support_ambulance: string;
  preparation: string;
  price: string;
  test_sku: string;
}